Changed Arenas to work in terms of Entities.

+++ src/com/dcdl/spear/Util.java
@@ -3,7 +3,8 @@ package com.dcdl.spear;
 import java.awt.Rectangle;
 
 public class Util {
-  public static Rectangle scaleRect(Rectangle rect, int scale) {
-    return new Rectangle(rect.x * scale, rect.y * scale, rect.width * scale, rect.height * scale);
+  public static Rectangle scaleRect(Rectangle rect, double scale) {
+    return new Rectangle((int) (rect.x * scale), (int) (rect.y * scale),
+        (int) (rect.width * scale), (int) (rect.height * scale));
   }
 }
